adding image files for better longevity

+++ update.sh
@@ -1,7 +1,12 @@
 #!/bin/zsh
 set -euo pipefail
-cd /home/spoopinatural/jagerbot
-git pull origin main
+cd /mnt/ssd/jagerbot
+PULL_OUTPUT=$(git pull origin main)
+
+if [[ "$PULL_OUTPUT" == *"Already up to date."* ]]; then
+    exit 0
+fi
+
 find . -type d -name "__pycache__" -exec rm -rf {} +
-/home/spoopinatural/jagerbot/venv/bin/pip install --upgrade -r requirements.txt
+/mnt/ssd/jagerbot/venv/bin/pip install --upgrade -r requirements.txt
 sudo systemctl restart jagerbot.service
